@@ -1,32 +1,32 @@
+/*
+   Copyright 2016 Yuki KAN
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
 /// <reference path="../typings/tsd.d.ts" />
 'use strict';
 
-import fs = require('fs');
+process.title = 'Mirakurun: Server';
+
+import mirakurun = require('./common');
 import uuid = require('node-uuid');
 
-import mirakurun = require('./index');
+// create initial uuid here
+// ...
+// ...
 
-var config: mirakurun.IServerConfig = {
-    id: uuid.v4(),
-
-    path: '',
-
-    tuners: [],
-    tunerGroups: []
-};
-
-if (process.env.MIRAKURUN_CONFIG_PATH) {
-    config.path = '/var/run/mirakurun.sock';
-    config.logLevel = 2;
-
-    if (fs.existsSync(process.env.MIRAKURUN_CONFIG_PATH) === true) {
-        config = require(process.env.MIRAKURUN_CONFIG_PATH);
-    }
-} else {
-    console.error('Note: running in debug mode for console.');
-
-    config.port = 40772;
-    config.logLevel = 3;
-}
-
-mirakurun.createServer(config);
+mirakurun.createServer({
+    server: mirakurun.config.getServer(),
+    tuner: mirakurun.config.getTuner(),
+    channels: mirakurun.config.getChannels()
+});
